Dev/rust llm clients (#309)
* Refactor LLM message handling to use structured content types

* message history update

* more client implementation

* more client implementation

* fix pre-commit check errors
