+++ trae_agent_rs/.gitignore
