+++ trae_agent_rs/core/src/config/mod.rs
@@ -10,7 +10,6 @@ pub struct ModelProvider {
     pub name: String,
     pub api_key: Option<String>,
     pub base_url: Option<String>,
-    pub api_version: Option<String>,
 }
 
 /// Model configuration for LLM clients
@@ -32,22 +31,16 @@ impl ModelProvider {
             name,
             api_key: None,
             base_url: None,
-            api_version: None,
         }
     }
-    
+
     pub fn with_api_key(mut self, api_key: String) -> Self {
         self.api_key = Some(api_key);
         self
     }
-    
+
     pub fn with_base_url(mut self, base_url: String) -> Self {
         self.base_url = Some(base_url);
-        self
-    }
-    
-    pub fn with_api_version(mut self, api_version: String) -> Self {
-        self.api_version = Some(api_version);
         self
     }
 }
@@ -64,27 +57,27 @@ impl ModelConfig {
             extra_headers: HashMap::new(),
         }
     }
-    
+
     pub fn with_temperature(mut self, temperature: f32) -> Self {
         self.temperature = Some(temperature);
         self
     }
-    
+
     pub fn with_top_p(mut self, top_p: f32) -> Self {
         self.top_p = Some(top_p);
         self
     }
-    
+
     pub fn with_max_tokens(mut self, max_tokens: u32) -> Self {
         self.max_tokens = Some(max_tokens);
         self
     }
-    
+
     pub fn with_max_retries(mut self, max_retries: u32) -> Self {
         self.max_retries = Some(max_retries);
         self
     }
-    
+
     pub fn with_extra_header(mut self, key: String, value: String) -> Self {
         self.extra_headers.insert(key, value);
         self
