+++ trae_agent_rs/core/src/lib.rs
