+++ trae_agent_rs/core/src/llm/openai_client.rs
@@ -13,7 +13,7 @@ use crate::{
         llm_provider::LLMProvider,
         error::{LLMError, LLMResult},
         retry_utils::{retry_with_backoff, RetryConfig},
-        LLMMessage, LLMResponse, LLMUsage, LLMStream, StreamChunk, FinishReason, ContentItem, MessageRole,
+        llm_basics::{LLMMessage, LLMResponse, LLMUsage, LLMStream, StreamChunk, FinishReason, ContentItem, MessageRole},
     },
     config::ModelConfig,
 };
@@ -138,7 +138,7 @@ impl OpenAIClient {
 
     async fn create_response(&self, request: OpenAIResponsesRequest) -> LLMResult<OpenAIResponse> {
         let url = format!("{}/responses", self.base_url);
-        
+
         let response = self
             .client
             .post(&url)
@@ -163,7 +163,7 @@ impl OpenAIClient {
 
     async fn create_stream_response(&self, request: OpenAIResponsesRequest) -> LLMResult<LLMStream> {
         let url = format!("{}/responses", self.base_url);
-        
+
         let response = self
             .client
             .post(&url)
@@ -184,7 +184,7 @@ impl OpenAIClient {
         }
 
         use futures::stream;
-        
+
         let text = response.text().await.map_err(LLMError::HttpError)?;
         let chunks: Vec<_> = text.lines().filter_map(|line| {
             if line.starts_with("data: ") {
@@ -201,7 +201,7 @@ impl OpenAIClient {
                 None
             }
         }).collect();
-        
+
         let stream = stream::iter(chunks);
 
         Ok(Box::pin(stream))
@@ -214,7 +214,7 @@ impl OpenAIClient {
                 if data == "[DONE]" {
                     return Ok(None);
                 }
-                
+
                 match serde_json::from_str::<OpenAIStreamResponse>(data) {
                     Ok(response) => {
                         if let Some(choice) = response.choices.first() {
@@ -228,7 +228,7 @@ impl OpenAIClient {
                                 model: response.model.clone(),
                                 tool_calls: choice.delta.tool_calls.as_ref().map(|calls| {
                                     calls.iter().map(|call| {
-                                        let arguments: HashMap<String, Value> = 
+                                        let arguments: HashMap<String, Value> =
                                             serde_json::from_str(&call.function.arguments)
                                                 .unwrap_or_default();
                                         ToolCall {
@@ -308,7 +308,7 @@ impl LLMProvider for OpenAIClient {
         reuse_history: Option<bool>,
     ) -> LLMResult<LLMResponse> {
         let parsed_messages = self.parse_messages(&messages);
-        
+
         let mut all_messages = Vec::new();
         if reuse_history.unwrap_or(true) {
             all_messages.extend(self.message_history.clone());
@@ -355,7 +355,7 @@ impl LLMProvider for OpenAIClient {
             for tool_call in response_tool_calls {
                 let arguments: HashMap<String, Value> = serde_json::from_str(&tool_call.function.arguments)
                     .unwrap_or_default();
-                
+
                 tool_calls.push(ToolCall {
                     name: tool_call.function.name.clone(),
                     call_id: tool_call.id.clone(),
@@ -408,10 +408,10 @@ impl LLMProvider for OpenAIClient {
         reuse_history: Option<bool>,
     ) -> LLMResult<LLMStream> {
         // Note: For streaming responses, chat history is not automatically updated.
-        // The caller should accumulate the complete response from the stream and 
+        // The caller should accumulate the complete response from the stream and
         // manually add it to chat history using set_chat_history() if needed.
         let parsed_messages = self.parse_messages(&messages);
-        
+
         let mut all_messages = Vec::new();
         if reuse_history.unwrap_or(true) {
             all_messages.extend(self.message_history.clone());
@@ -440,10 +440,5 @@ impl LLMProvider for OpenAIClient {
 
     fn get_provider_name(&self) -> &str {
         "openai"
-    }
-
-    fn supports_tool_calling(&self, model_name: &str) -> bool {
-        let model_lower = model_name.to_lowercase();
-        model_lower.contains("gpt-4") || model_lower.contains("gpt-3.5-turbo")
     }
 }
